Create a SharedKmer class
This class keeps track of which files share the same kmer.

During an analysis, we only create a collection of SharedKmers and only
afterwards we build the Intersection > Fragment > Match structure. This
allows us to keep the most basic representation of plagiarism and to
perform filtering on the level of kmers (e.g. throw away if >50% of
files have the same kmer).